LayoutTests/ietestcenter/Javascript/15.2.3.10-2-expected.txt

15.2.3.10-2 (TestCases/chapter15/15.2/15.2.3/15.2.3.10/15.2.3.10-2.js) Object.preventExtensions returns its arguments after setting its extensible property to false

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL ES5Harness.preconditionPassed should be true (of type boolean). Was undefined (of type undefined).
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
